+++ server/odc-service/src/main/java/com/oceanbase/odc/service/resource/ResourceManagerService.java
@@ -15,9 +15,7 @@
  */
 package com.oceanbase.odc.service.resource;
 
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
@@ -25,7 +23,6 @@ import org.apache.commons.collections4.CollectionUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.oceanbase.odc.service.resource.builder.ResourceOperatorBuilder;
 import com.oceanbase.odc.service.resource.operator.ResourceOperator;
 
 import lombok.NonNull;
@@ -43,10 +40,14 @@ import lombok.extern.slf4j.Slf4j;
 public class ResourceManagerService {
 
     @Autowired(required = false)
-    private List<ResourceOperatorBuilder<?, ?>> resourceOperatorBuilders;
+    private List<ResourceOperator<?, ?>> resourceOperators;
 
     public <T> T create(@NonNull T config) throws Exception {
         return getResourceOperatorByConfig(config).create(config);
+    }
+
+    public <T> Object getKey(@NonNull T config) throws Exception {
+        return getResourceOperatorByConfig(config).getKey(config);
     }
 
     public <T> Optional<T> query(@NonNull Object key, @NonNull Class<T> clazz) throws Exception {
@@ -61,12 +62,6 @@ public class ResourceManagerService {
         return getResourceOperatorByConfig(clazz).list();
     }
 
-    protected Map<String, Object> getParameterForOperatorBuilder(@NonNull ResourceOperatorBuilder<?, ?> builder) {
-        Map<String, Object> parameter = new HashMap<>();
-        parameter.put(ResourceConstants.DEFAULT_NAMESPACE_PARAMETER_KEY, "shanlu");
-        return parameter;
-    }
-
     @SuppressWarnings("all")
     protected <T, ID> ResourceOperator<T, ID> getResourceOperatorByConfig(T config) {
         return (ResourceOperator<T, ID>) getResourceOperatorByConfig(config.getClass());
@@ -74,15 +69,14 @@ public class ResourceManagerService {
 
     @SuppressWarnings("all")
     protected <T, ID> ResourceOperator<T, ID> getResourceOperatorByConfig(Class<T> clazz) {
-        List<ResourceOperatorBuilder<?, ?>> builders = this.resourceOperatorBuilders.stream()
+        List<ResourceOperator<?, ?>> builders = this.resourceOperators.stream()
                 .filter(builder -> builder.supports(clazz)).collect(Collectors.toList());
         if (CollectionUtils.isEmpty(builders)) {
             throw new IllegalArgumentException("No builder found for config " + clazz);
         } else if (builders.size() != 1) {
             throw new IllegalStateException("There are more than one builder for the config " + clazz);
         }
-        ResourceOperatorBuilder<T, ID> builder = (ResourceOperatorBuilder<T, ID>) builders.get(0);
-        return builder.build(getParameterForOperatorBuilder(builder));
+        return (ResourceOperator<T, ID>) builders.get(0);
     }
 
 }
