+++ server/odc-service/src/main/java/com/oceanbase/odc/service/resource/operator/BaseNativeK8sResourceOperator.java
@@ -15,14 +15,33 @@
  */
 package com.oceanbase.odc.service.resource.operator;
 
+import java.io.ByteArrayInputStream;
+import java.io.IOException;
+import java.io.InputStreamReader;
+import java.io.Reader;
 import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
+import java.util.concurrent.TimeUnit;
 
+import javax.annotation.PostConstruct;
+
+import org.apache.commons.lang3.StringUtils;
+import org.springframework.beans.factory.annotation.Autowired;
+
+import com.oceanbase.odc.common.util.EncodeUtils;
+import com.oceanbase.odc.core.shared.Verify;
 import com.oceanbase.odc.service.resource.model.NativeK8sResourceKey;
+import com.oceanbase.odc.service.task.config.K8sProperties;
+import com.oceanbase.odc.service.task.config.TaskFrameworkProperties;
 
 import cn.hutool.core.collection.CollectionUtil;
 import io.kubernetes.client.common.KubernetesObject;
+import io.kubernetes.client.openapi.ApiClient;
+import io.kubernetes.client.openapi.Configuration;
+import io.kubernetes.client.util.ClientBuilder;
+import io.kubernetes.client.util.Config;
+import io.kubernetes.client.util.KubeConfig;
 import lombok.NonNull;
 
 /**
@@ -35,10 +54,51 @@ import lombok.NonNull;
 public abstract class BaseNativeK8sResourceOperator<T extends KubernetesObject>
         implements ResourceOperator<T, NativeK8sResourceKey> {
 
-    protected final String defaultNamespace;
+    private static final long TIMEOUT_MILLS = 60000;
+    protected final String defaultNamespace = "shanlu";
+    private final Object lock = new Object();
+    private volatile boolean apiClientAvailable = false;
+    private volatile boolean apiClientSet = false;
+    @Autowired
+    private TaskFrameworkProperties taskFrameworkProperties;
 
-    public BaseNativeK8sResourceOperator(@NonNull String defaultNamespace) {
-        this.defaultNamespace = defaultNamespace;
+    @PostConstruct
+    public void setUp() throws IOException {
+        if (this.apiClientSet) {
+            return;
+        }
+        synchronized (this.lock) {
+            if (this.apiClientSet) {
+                return;
+            }
+            this.apiClientSet = true;
+            K8sProperties properties = this.taskFrameworkProperties.getK8sProperties();
+            if (properties == null) {
+                return;
+            }
+            ApiClient apiClient;
+            if (StringUtils.isNotBlank(properties.getKubeConfig())) {
+                byte[] config = EncodeUtils.base64DecodeFromString(properties.getKubeConfig());
+                Verify.notNull(config, "kube config");
+                try (Reader reader = new InputStreamReader(new ByteArrayInputStream(config))) {
+                    KubeConfig kubeConfig = KubeConfig.loadKubeConfig(reader);
+                    apiClient = ClientBuilder.kubeconfig(kubeConfig).build();
+                }
+            } else if (StringUtils.isNotBlank(properties.getKubeUrl())) {
+                apiClient = Config.defaultClient().setBasePath(properties.getKubeUrl());
+            } else {
+                return;
+            }
+            apiClient.setHttpClient(apiClient
+                    .getHttpClient()
+                    .newBuilder()
+                    .readTimeout(TIMEOUT_MILLS, TimeUnit.MILLISECONDS)
+                    .connectTimeout(TIMEOUT_MILLS, TimeUnit.MILLISECONDS)
+                    .pingInterval(1, TimeUnit.MINUTES)
+                    .build());
+            Configuration.setDefaultApiClient(apiClient);
+            this.apiClientAvailable = true;
+        }
     }
 
     @Override
@@ -49,5 +109,12 @@ public abstract class BaseNativeK8sResourceOperator<T extends KubernetesObject>
         }
         return list.stream().filter(t -> Objects.equals(getKey(t), key)).findFirst();
     }
+
+    @Override
+    public boolean supports(@NonNull Class<?> clazz) {
+        return this.apiClientAvailable && doSupports(clazz);
+    }
+
+    protected abstract boolean doSupports(@NonNull Class<?> clazz);
 
 }
